% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasissa-lai.R
\name{alternating_completion}
\alias{alternating_completion}
\title{Tasissa-Lai Algorithm}
\usage{
alternating_completion(D, Selected, opts, lsopts)
}
\arguments{
\item{D}{partial distance}

\item{Selected}{selected missing}

\item{opts}{options}

\item{lsopts}{other options}
}
\description{
CONVERTED FROM:
Abiy Tasissa and Rongjie Lai, "Exact Reconstruction of Euclidean Distance Geometry Problem Using Low-rank Matrix Completion," in IEEE Transactions on Information Theory, 2018. doi: 10.1109/TIT.2018.2881749
https://github.com/abiy-tasissa/Nonconvex-Euclidean-Distance-Geometry-Problem-Solver

TODO: needs testing!
}
\note{
:
Original Note
--------------------------------------------------------------------------
This script is a numerical solution of the Euclidean distance geometry
problem (EDG). Formally, consider a set of n points where partial
inter-point distance information is provided. The goal of EDG is to find
the coordinate of the points given this partial information.
--------------------------------------------------------------------------
The n points usually lie in a low dimensional space of size r << n, low
rank. With this, the EDG problem can be set as low-rank completion problem
which can be solved via nuclear norm minimization. We recover the Gram
matrix, the inner product matrix, and follow classical MDS to recover the
coordinates. For details, see the associated paper below.
--------------------------------------------------------------------------
Tasissa, Abiy, and Rongjie Lai."Exact Reconstruction of Euclidean Distance
Geometry Problem Using Low-rank Matrix Completion." arXiv preprint
arXiv:1804.04310 (2018).
--------------------------------------------------------------------------
The gram matrix is psd so nuclear norm minimization equates to trace.
Our algorithm uses the Augmented Lagrangian framework to find the Gram
matrix.
--------------------------------------------------------------------------
We assume the partial inter-distance information comes from a uniform
random sample. The information is assumed to be exact.
Dist is the full distance matrix. Weight is a binary matrix informing
whether a given entry of Dist is chosen or not.
--------------------------------------------------------------------------
Rongjie Lai, Abiy Tasissa
--------------------------------------------------------------------------
}
