% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavaud.R
\name{disputedness}
\alias{disputedness}
\title{Compute column disputedness}
\usage{
disputedness(M, f = NULL)
}
\arguments{
\item{M}{a matrix containing NAs}

\item{f}{an array containing weights}
}
\value{
a named array containing the column disputedness
}
\description{
Compute column disputedness
}
\examples{
M <- matrix(c(1, 2, NA, NA, 4, 19, 0, NA, 0), nrow = 3)
disputedness(M)
}
