% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavaud.R
\name{dissimilarity.disputedness}
\alias{dissimilarity.disputedness}
\title{Compute the disputedness dissimilarity}
\usage{
dissimilarity.disputedness(M, f, disputedness)
}
\arguments{
\item{M}{a matrix containing NAs}

\item{f}{a named weight array}

\item{disputedness}{an array containing column disputedness}
}
\value{
a dissimilarity renormalised by disputedness
}
\description{
Compute the disputedness dissimilarity
}
\examples{
M <- matrix(c(1, 2, NA, NA, 4, 19, 0, NA, 0), nrow = 3)
f <- rowSums(M, na.rm = TRUE) / sum(M, na.rm = TRUE)
d <- disputedness(M)
dissimilarity.disputedness(M, f, d)
}
