% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavaud.R
\name{dissimilarity.L1}
\alias{dissimilarity.L1}
\title{Compute the L1 dissimilarity}
\usage{
dissimilarity.L1(M)
}
\arguments{
\item{M}{a matrix containing NAs}
}
\value{
a dissimilarity renormalised by disputedness
}
\description{
Compute the L1 dissimilarity
}
\examples{
M <- matrix(c(1, 2, NA, NA, 4, 19, 0, NA, 0), nrow = 3)
f <- rowSums(M, na.rm = TRUE) / sum(M, na.rm = TRUE)
dissimilarity.L1(M)
}
