% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavaud.R
\name{dissimilarity.regression_estimation}
\alias{dissimilarity.regression_estimation}
\title{Compute the dissimilarity estimation}
\usage{
dissimilarity.regression_estimation(D)
}
\arguments{
\item{D}{a dissimilarity}
}
\description{
Compute the dissimilarity estimation
}
\examples{
M <- matrix(c(1, 2, NA, NA, 4, 19, 0, NA, 0), nrow = 3)
f <- rowSums(M, na.rm = TRUE) / sum(M, na.rm = TRUE)
d <- disputedness(M)
D <- dissimilarity.disputedness(M, f, d)
D_estim <- dissimilarity.regression_estimation(D)
}
