% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavaud.R
\name{dissimilarity_disputedness}
\alias{dissimilarity_disputedness}
\title{dissimilarity_disputedness}
\usage{
dissimilarity_disputedness(M, f, disp)
}
\arguments{
\item{M}{a matrix containing NAs}

\item{f}{a named weight array}

\item{disp}{an array containing column disputedness}
}
\value{
a dissimilarity renormalised by disputedness
}
\description{
dissimilarity_disputedness
}
