% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavaud.R
\name{dissimilarity_final}
\alias{dissimilarity_final}
\title{dissimilarity_final}
\usage{
dissimilarity_final(D_estim, D)
}
\arguments{
\item{D_estim}{an estimated dissimilarity}

\item{D}{a dissimilarity}
}
\description{
dissimilarity_final
}
