% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavaud.R
\name{dissimilarity_L1}
\alias{dissimilarity_L1}
\title{dissimilarity_L1}
\usage{
dissimilarity_L1(M)
}
\arguments{
\item{M}{a matrix containing NAs}
}
\value{
a dissimilarity renormalised by disputedness
}
\description{
dissimilarity_L1
}
