% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavaud.R
\name{dissimilarity_regression_estimation}
\alias{dissimilarity_regression_estimation}
\title{dissimilarity_regression_estimation}
\usage{
dissimilarity_regression_estimation(D)
}
\arguments{
\item{D}{a dissimilarity}
}
\description{
dissimilarity_regression_estimation
}
