% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavaud.R
\name{dissimilarity_tilde_estimation}
\alias{dissimilarity_tilde_estimation}
\title{dissimilarity_tilde_estimation}
\usage{
dissimilarity_tilde_estimation(D, f = NULL)
}
\arguments{
\item{D}{a dissimilarity}

\item{f}{a named weight array}
}
\value{
a dissimilarity estimation
}
\description{
dissimilarity_tilde_estimation
}
