% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavaud.R
\name{estimate.distance}
\alias{estimate.distance}
\title{Compute the distance estimation}
\usage{
estimate.distance(M, f)
}
\arguments{
\item{M}{a matrix}

\item{f}{a named weight array}
}
\value{
a dissimilarity estimation
}
\description{
Compute the distance estimation
}
\examples{
M <- matrix(c(1, 2, NA, NA, 4, 19, 0, NA, 0), nrow = 3)
f <- rowSums(M, na.rm = TRUE) / sum(M, na.rm = TRUE)
D <- estimate.distance(M, f)
}
