% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavaud.R
\name{estimate_distance}
\alias{estimate_distance}
\title{estimate_distance}
\usage{
estimate_distance(M, f, disp = NULL)
}
\arguments{
\item{M}{a matrix}

\item{f}{a named weight array}

\item{disp}{disputedness}
}
\value{
a dissimilarity estimation
}
\description{
estimate_distance
}
