% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalutation.R
\name{NA_dist_random}
\alias{NA_dist_random}
\title{NA_dist_random}
\usage{
NA_dist_random(M, percentage)
}
\arguments{
\item{M}{a symmetric matrix}

\item{percentage}{the percentage of elements to replace by NA}
}
\description{
Randomly sets a percentage of elements of a symmetric matrix to NA. diag being excluded from replacements.
}
