\name{RNcEDGPS-package}
\alias{RNcEDGPS-package}
\alias{RNcEDGPS}
\docType{package}
\title{
\packageTitle{RNcEDGPS}
}
\description{
\packageDescription{RNcEDGPS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RNcEDGPS}
\packageIndices{RNcEDGPS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RNcEDGPS}

Maintainer: \packageMaintainer{RNcEDGPS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
