% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalutation.R
\name{sphere_vertices}
\alias{sphere_vertices}
\title{sphere_vertices}
\usage{
sphere_vertices(xc = 0, yc = 0, zc = 0, r = 1, lats = 16L, longs = 16L)
}
\arguments{
\item{xc}{x cooddinate for the center of the sphere}

\item{yc}{y cooddinate for the center of the sphere}

\item{zc}{z cooddinate for the center of the sphere}

\item{r}{radius}

\item{lats}{number of latitudes}

\item{longs}{number of longitudes}
}
\description{
sphere_vertices
}
