% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavaud.R
\name{validity.matrix}
\alias{validity.matrix}
\title{Compute validity matrix}
\usage{
validity.matrix(M)
}
\arguments{
\item{M}{a matrix containing NAs}
}
\value{
The validity matrix containing 1 if the element is a NA 0 else.
}
\description{
Compute validity matrix
}
\examples{
M <- matrix(c(1, 2, NA, NA, 4, 19, 0, NA, 0), nrow = 3)
validity.matrix(M)
}
