% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavaud.R
\name{validity_matrix}
\alias{validity_matrix}
\title{validity_matrix}
\usage{
validity_matrix(M)
}
\arguments{
\item{M}{a matrix containing NAs}
}
\value{
The validity matrix containing 1 if the element is a NA 0 else.
}
\description{
validity_matrix
}
