% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavaud.R
\name{validity_weight}
\alias{validity_weight}
\title{validity_weight}
\usage{
validity_weight(M)
}
\arguments{
\item{M}{a matrix containing NAs}
}
\value{
a named array containing weights
}
\description{
validity_weight
}
