% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavaud.R
\name{disputedness}
\alias{disputedness}
\title{disputedness}
\usage{
disputedness(M, f = NULL)
}
\arguments{
\item{M}{a matrix containing NAs}

\item{f}{an array containing weights}
}
\value{
a named array containing the column disputedness
}
\description{
disputedness
}
