% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavaud.R
\name{dissimilarity_L1}
\alias{dissimilarity_L1}
\title{Compute the L1 dissimilarity}
\usage{
dissimilarity_L1(M)
}
\arguments{
\item{M}{a matrix containing NAs}
}
\value{
a dissimilarity renormalised by disputedness
}
\description{
Compute the L1 dissimilarity
}
\examples{
M <- matrix(c(1, 2, NA, NA, 4, 19, 0, NA, 0), nrow = 3)
dissimilarity_L1(M)
}
