% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavaud.R
\name{dissimilarity_tilde_estimation}
\alias{dissimilarity_tilde_estimation}
\title{Compute the dissimilarity estimation}
\usage{
dissimilarity_tilde_estimation(D, f = NULL)
}
\arguments{
\item{D}{a dissimilarity}

\item{f}{a named weight array}
}
\value{
a dissimilarity estimation
}
\description{
Compute the dissimilarity estimation
}
\examples{
M <- matrix(c(1, 2, NA, NA, 4, 19, 0, NA, 0), nrow = 3)
f <- rowSums(M, na.rm = TRUE) / sum(M, na.rm = TRUE)
disp <- disputedness(M)
D <- dissimilarity_disputedness(M, f, disp)
dissimilarity_tilde_estimation(D, f)
}
