% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavaud.R
\name{validity_weight}
\alias{validity_weight}
\title{Compute validity weight}
\usage{
validity_weight(M)
}
\arguments{
\item{M}{a matrix containing NAs}
}
\value{
a named array containing weights
}
\description{
Compute validity weight
}
\examples{
M <- matrix(c(1, 2, NA, NA, 4, 19, 0, NA, 0), nrow = 3)
validity_weight(M)
}
